Final_Presentation/DOPL/DOPL/requirements.txt
amply @ file:///home/conda/feedstock_root/build_artifacts/amply_1687675480808/work
antlr4-python3-runtime==4.9.3
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1720457958366/work/certifi
charset-normalizer==3.3.2
click==8.1.7
cloudpickle @ file:///home/conda/feedstock_root/build_artifacts/cloudpickle_1697464713350/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1712429964190/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
docker-pycreds==0.4.0
docutils @ file:///home/conda/feedstock_root/build_artifacts/docutils_1713930378730/work
Farama-Notifications @ file:///home/conda/feedstock_root/build_artifacts/farama-notifications_1684258120643/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1719088281970/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1712344641187/work
fsspec @ file:///home/conda/feedstock_root/build_artifacts/fsspec_1719514913127/work
gitdb==4.0.11
GitPython==3.1.43
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1715527354984/work
gurobipy==11.0.1
gymnasium @ file:///Users/runner/miniforge3/conda-bld/gymnasium-split_1712294473351/work
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.7
jax-jumpy @ file:///home/conda/feedstock_root/build_artifacts/jax-jumpy_1679161251636/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1715127149914/work
joblib==1.4.2
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1695379982481/work
llvmlite==0.43.0
markovianbandit-pkg==0.4
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1706900061465/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1712606004214/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
munkres==1.1.4
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1712540363324/work
numba==0.60.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1707225416355/work/dist/numpy-1.26.4-cp310-cp310-macosx_11_0_arm64.whl#sha256=14c628943f67e9c10d22faa9c1cdcbf69c3192e88f2f597d03b77a252e589567
omegaconf==2.3.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1710075952259/work
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1712782129530/work
pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1712154790137/work
platformdirs==4.2.2
ply @ file:///home/conda/feedstock_root/build_artifacts/ply_1712242996588/work
protobuf==4.25.3
psutil==5.9.8
PuLP @ file:///Users/runner/miniforge3/conda-bld/pulp_1705065039439/work
Pyomo @ file:///Users/runner/miniforge3/conda-bld/pyomo_1715275060872/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1709721012883/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1706886791323/work
PyYAML==6.0.1
requests==2.31.0
scipy==1.13.0
sentry-sdk==2.2.0
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1721004824012/work
torch @ file:///Users/runner/miniforge3/conda-bld/libtorch_1719361051023/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1708363219318/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1714854870413/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1712329955671/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1707747584337/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1695848003431/work
urllib3==2.2.1
wandb==0.17.0
